requirements.txt
numpy==1.24.1
opencv_python==4.8.1.78
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.3.2
six==1.16.0
torch==2.1.0+cu121
torchvision==0.16.0+cu121
tqdm==4.66.1
transformers==4.36.2
